Fixing rustup invocations
I couldn't figure out how to test this locally without setting up docker
or a vm, so I tested with CI (sigh). Apparently my local machine has
llvm-tools-preview installed for nightly and stable, and the old command
was installing the stable version.

On a fresh machine, you can't execute the command as it was written --
it complains that the toolchain isn't installed yet. I thought this
one-liner in one of the blog posts I found was to just do both things in
one go.

This hopefully will make coverage work again.